requirements.txt
python-telegram-bot==13.7
python-dotenv=0.19.0
google-cloud-dialogflow==2.7.1
environs==9.3.3
vk-api==11.9.4
